Move fixtures to `hss_fixtures.py`
Also, make the fixture classes pure mixins, and handle the
initialization in the `HssPgTestBase` instead.
